reafactor: move profile into components
